fix(Issue #41): Implemented a possible fix for the furnace GUI issue.
I noticed that the furnaceBlockPos was not being saved or loaded correctly, but it also didn't seem implemented right. When I had multiple blacksmiths down, I was only able to access a single furnace. Using this new system, the furnace blocks within each Blacksmith are scanned, then the position is returned. This prevents needing to save the furnace position in the case where it fails. This fix will also make existing broken worlds become functional again.