Resolved PR #42 comments
I have removed all references to the original 'getFurnace()' method, and replaced it with 'getFurnacePos()'.

I have also changed getFurnacePos() to simply return a cached list of furnaces. The functionality has been moved to 'findFurnaces()', which is called in 'init()', and if the furnace list was null when performing the DailyTask. This allows the furnace list to be initialized so long as it is uninitialized, and updated when a new building is placed.

I have modified the solution to return a list of furnaces for each building (in all cases). Now, when clicking the furnace icon, the amount of furnaces available to use is equal to the number of upgrades at each building, times 2.

Also simplified the method of detecting if a block is a furnace. I reverse engineered too hard.