Remove improperly implemented furnace limit
I had tried to implement a limit on the number of furnaces per building. The goal was to have level 1 be 2 furnaces, level 2 be 4 furnaces (max) etc. I was using a variable that didn't seem to exist. From what I can see, there's no direct reference to the upgrade number, so maybe I'll just leave this for now...